map helper was added

+++ HSE_Transport1/DataModels/Bus.cs
@@ -18,7 +18,7 @@ namespace HSE_Transport1
         public DateTime DepartureTime { get; set; }
         public string DeparturePlace { get; set; }
         public string ArrivalPlace { get; set; }
-        public int JourneyDuration { get; set; }
+        public double JourneyDuration { get; set; }
         public string Occupancy { get; set; }
         public bool Notify { get; set; }
         public string Day { get; set; }
